I moved crossover operators to a separate file crossover_operators.py from the genetic_algorithm.py file
My aim is to create an organised library with the ability to construct Pipelines in a main file, thus having different kinds of operators in separate locations and importing them - this way it's easier for User to use a custom operator, defined by them, as they'll know where to put it.
